+++ src/app/view-config-versions.js
@@ -12,6 +12,7 @@ import configSchema1_0_7 from '../schemas/config-1.0.7.schema.json';
 import configSchema1_0_8 from '../schemas/config-1.0.8.schema.json';
 import configSchema1_0_9 from '../schemas/config-1.0.9.schema.json';
 import configSchema1_0_10 from '../schemas/config-1.0.10.schema.json';
+import configSchema1_0_11 from '../schemas/config-1.0.11.schema.json';
 // TODO
 import configSchema1_0_15 from '../schemas/config-1.0.15.schema.json';
 import cellSetsSchema from '../schemas/cell-sets.schema.json';
@@ -28,6 +29,7 @@ import {
   upgradeFrom1_0_7,
   upgradeFrom1_0_8,
   upgradeFrom1_0_9,
+  upgradeFrom1_0_10,
   // TODO
   upgradeFrom1_0_14,
 } from './view-config-upgraders';
@@ -50,8 +52,8 @@ export const SCHEMA_HANDLERS = {
   '1.0.7': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_7), upgradeFrom1_0_7],
   '1.0.8': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_8), upgradeFrom1_0_8],
   '1.0.9': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_9), upgradeFrom1_0_9],
-  '1.0.10': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_10), upgradeFrom1_0_14], // TODO
+  '1.0.10': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_10), upgradeFrom1_0_10],
+  '1.0.11': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_11), upgradeFrom1_0_14], // TODO
   // TODO
   '1.0.15': [new Ajv().addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_15), null],
-
 };